requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
flask-marshmallow==1.2.1
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
marshmallow==3.21.3
marshmallow-sqlalchemy==1.0.0
packaging==24.0
PyJWT==2.8.0
PyMySQL==1.1.1
SQLAlchemy==2.0.30
typing_extensions==4.12.2
Werkzeug==3.0.3
